chore #8: 불필요한 로직 삭제
