+++ trecord/src/apis/Alarm/getAlarm.ts
@@ -1,6 +1,6 @@
 import { UseQueryResult, useQuery } from '@tanstack/react-query';
 import { http } from '../_http';
-import { GetAlarmResponse, GetAllAlarm } from '@/types/alarm';
+import { GetAlarm } from '@/types/alarm';
 import ALARM_API_KEY from './constants';
 import { Page } from '@/types';
 interface Props {
@@ -11,19 +11,19 @@ interface Props {
 const getAlarm = async ({
   pageCount,
   alarmType,
-}: Props): Promise<Page<GetAllAlarm>> => {
+}: Props): Promise<Page<GetAlarm>> => {
   const url = `/api/v1/notifications${
     alarmType === 'ALL' ? `` : `/type/${alarmType}`
   }?page=0&size=${pageCount}`;
 
-  const response: Page<GetAllAlarm> = await http.authGet(url);
+  const response: Page<GetAlarm> = await http.authGet(url);
   return response;
 };
 
 const useGetAlarm = ({
   pageCount,
   alarmType,
-}: Props): UseQueryResult<Page<GetAllAlarm>> => {
+}: Props): UseQueryResult<Page<GetAlarm>> => {
   return useQuery(
     [ALARM_API_KEY.ALL_ALARM, { pageCount, alarmType }],
     () => getAlarm({ pageCount, alarmType }),
