+++ trecord/src/apis/Comment/getNewComment.ts
@@ -1,6 +1,6 @@
 import { UseQueryResult, useQuery } from '@tanstack/react-query';
 import { http } from '../_http';
-import { GetCommentProps } from '@/types/comment';
+import { GetComment } from '@/types/comment';
 import COMMENT_API_KEY from './constants';
 import { Page } from '@/types';
 
@@ -11,16 +11,16 @@ interface Props {
 export const getNewComment = async ({
   recordId,
   page,
-}: Props): Promise<Page<GetCommentProps>> => {
+}: Props): Promise<Page<GetComment>> => {
   const url = `/api/v1/records/${recordId}/comments?page=0&size=${page}`;
-  const response: Page<GetCommentProps> = await http.get(url);
+  const response: Page<GetComment> = await http.get(url);
   return response;
 };
 
 const useGetNewComment = ({
   recordId,
   page,
-}: Props): UseQueryResult<Page<GetCommentProps>> => {
+}: Props): UseQueryResult<Page<GetComment>> => {
   return useQuery(
     [COMMENT_API_KEY.NEW_COMMENT, { record_id: recordId, page }],
     () => getNewComment({ recordId, page }),
