+++ trecord/src/apis/Comment/getReplyComment.ts
@@ -1,6 +1,6 @@
 import { UseQueryResult, useQuery } from '@tanstack/react-query';
 import { http } from '../_http';
-import { GetReplyCommentProps } from '@/types/comment';
+import { GetReplyComment } from '@/types/comment';
 import COMMENT_API_KEY from './constants';
 import { Page } from '@/types';
 
@@ -11,16 +11,16 @@ interface Props {
 export const getReplyComment = async ({
   commentId,
   pageCount,
-}: Props): Promise<Page<GetReplyCommentProps>> => {
+}: Props): Promise<Page<GetReplyComment>> => {
   const url = `/api/v1/comments/${commentId}/replies?page=0&size=${pageCount}`;
-  const response: Page<GetReplyCommentProps> = await http.get(url);
+  const response: Page<GetReplyComment> = await http.get(url);
   return response;
 };
 
 const useGetReplyComment = ({
   commentId,
   pageCount,
-}: Props): UseQueryResult<Page<GetReplyCommentProps>> => {
+}: Props): UseQueryResult<Page<GetReplyComment>> => {
   return useQuery(
     [
       COMMENT_API_KEY.REPLY_COMMENT,
