+++ trecord/src/components/Alarm/AlarmList/style.ts
