--- trecord/src/components/FeedDetail/FeedCategoryList/index.tsx
@@ -1,10 +1,0 @@
-import * as S from './style';
-export const FeedCategoryList = () => {
-  return (
-    <S.Layout>
-      <S.EditBtnBox>수정하기</S.EditBtnBox>
-      <S.CancelBtnBox>삭제하기</S.CancelBtnBox>
-      <S.ShareBtnBox>공유하기</S.ShareBtnBox>
-    </S.Layout>
-  );
-};
