+++ trecord/src/components/MypageComment/MypageCommentList.tsx
@@ -1,13 +1,14 @@
-import { GetMypageCommentResponse } from '@/types/comment';
 import { Icon } from '@components/common/Icon';
 import * as S from './style';
 import { Fragment } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { replaceDate } from '@/utils/replaceDate';
 import Pagination from '@components/common/Pagination';
+import { Page } from '@/types';
+import { GetMypageComment } from '@/types/mypage';
 interface Props {
   onModalActive: (id: number) => void;
-  commentData: GetMypageCommentResponse;
+  commentData: Page<GetMypageComment>;
   onPageCount: () => void;
 }
 export const MypageCommentList = ({
