+++ trecord/src/components/common/Modal/ModalBody/InputContainerFallback.tsx
@@ -1,4 +1,4 @@
-import Skeleton from '@components/common/skeleton';
+import Skeleton from '@components/common/Skeleton';
 
 const InputContainerFallback = ({ inputValue }: { inputValue: string }) => {
   return (
