--- trecord/src/components/common/NewBtn/NewRecordBtn/index.tsx
@@ -1,15 +1,0 @@
-import { Icon } from '@components/common/Icon';
-import * as S from './style';
-import icons from '@/assets/index';
-interface NewBtnProps {
-  type: keyof typeof icons;
-  iconWidth: number;
-  onClick?: () => void;
-}
-export const NewRecordBtn = ({ type, iconWidth, onClick }: NewBtnProps) => {
-  return (
-    <S.Layout>
-      <Icon iconType={type} width={iconWidth} onClick={onClick} />
-    </S.Layout>
-  );
-};
