+++ trecord/src/pages/RecordDetail/index.tsx
@@ -9,7 +9,7 @@ import { Icon } from '@components/common/Icon';
 import { colorStyles } from '@/styles/color';
 import Modal from '@components/common/Modal';
 import { TabBarRecord } from '@components/common/TabBar/TabBarRecord';
-import Skeleton from '@components/common/skeleton';
+import Skeleton from '@components/common/Skeleton';
 import ShareModalBody from '@components/common/Modal/ModalBody/ShareModalBody';
 
 import SelectButton from '@components/common/button/SelectButton';
